packages/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
mem0ai==0.1.0
chromadb==0.4.18
langchain==0.0.350
langchain-openai==0.0.2
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
sse-starlette==1.8.2
PyPDF2==3.0.1
python-docx==1.1.0
Pillow>=10.2.0
sentence-transformers==2.2.2
numpy>=1.26.0
tiktoken==0.5.2
supabase>=2.0.0
